Update, add, many features from suggestions.

+++ src/main/java/org/kbot/UtilityCommands/Purge.java
@@ -11,6 +11,11 @@ public class Purge {
         String num = interaction.getArgumentStringValueByIndex(0).orElse("Invalid");
         int deleteAmt = Integer.parseInt(num);
         String returnMessage;
+        if (deleteAmt < 1) {
+            returnMessage = "Purge amount must be greater than 0";
+            interaction.createImmediateResponder().setContent(returnMessage).setFlags(MessageFlag.EPHEMERAL).respond();
+            return;
+        }
         if (CheckPermission.checkPermission(interaction, PermissionType.MANAGE_MESSAGES)) {
             try {
                 interaction.getChannel().get().getMessages(deleteAmt).get().deleteAll();
